Add new hyper/race pipeline builder
This is the part that kicks off parallel workers. It does so using
non-blocking constructs, so as not to exhaust the thread pool and to
play nice with other tasks ongoing in the pool. This also means we'll
have nice behavior if a parallel worker does a `react` or `await` (so,
non-blocking under 6.d.PREVIEW). Support for rebatchers still to come.