Clear up `for` and `hyper|race|lazy for` semantics
A plain `for` loop will now always mean serial execution, to avoid
surprises (for example, in `for some-module-code() { }` it is far
preferable to know that the module code can't suddenly cause the `for`
loop's body to run on different threads by changing its return type!)
The `hyper for foo() { }` and `race for foo() { }` constructs will run
the loop body in parallel under the `hyper` or `race` paradigm (these
are identical in sink context). Finally, `lazy for @foo { }` is fixed
to actually lazily execute the loop body when data is needed; this was
broken before now.