Support rw integer arguments of JIT compiled native subs
After the call to the native function, we need to read back the changed
values from the args buffer. We can't use the param_* ops there because
we don't set up a call frame for the native call. Thus the param_* ops
would read the HLL sub's arguments instead of the native function's.
We use the new getarg_i op instead to get the changed value and assign
it to the lexicalref for the rw parameter.