Add the \ on volumes in !canon-cat
The regex being used in !canon-cat does actually parse Volumes
with and wihtout \ in the end of a volume. When the method
concatenats the @rest it does now always check if the \ is missing.

This leads to a spec test violation at S32-io/io-spec-win.t.
See also:
https://github.com/perl6/roast/commit/8ee0a0a6a5a3ddd7bdef0212b29ffc717742cc2b#diff-29c343608a909127cb67b1f0dfb6caefR319

IMHO the test in wrong, because when you specify :volume<C:>, :$basename
I am assuming the $basename is located in C:. So it should be something
like "C:\$basename".