Fix JSONizing exceptions, spotted by Garland_g[m]++
When running with RAKUDO_EXCEPTIONS_HANDLER=JSON, any Exception that
contains another Exception as one of the attributes, would throw.  And
since we apparently don't have a CATCHer there, would give a very LTA
error message.  Fixed by moving the JSONizing of an Exception to
Rakudo::Internals::JSON, so it can recursively handle itself.  And
adapt the exception handling in Exception accordingly.