A more HLL approach to handling exceptions in END
The original code was sorta cargo-culted from NQP at the time.
Having it in HLL code means we can also handle exceptions at this
point in time in HLL code.  Also it seemed like a good idea to not
have one END block's exception blow it for all of the other ones.
Finally, give a summary of execution errors that occurred when done.