Expose new R:I:While/Until in .head/.tail
- .head taking a Callable:D: block should return trueish while values
  should be passed on.  The first falsish value will end the iterator.
- .tail taking a Callable:D: block should return falsish while values
  are being skipped.  The first trueish value will start passing on
  values until iterator is exhausted.
- no breakage in spectest seen.

Please note that I feel these semantics are a bit in conflict with the
"* - x" semantics meaning "so many from end".  So this would be the first
case where a WhateverCode has a different semantic from a generic Callable.
Perhaps we should expose this functionality in separate .while/.until
methods.  Please discuss :-)