Pass typenames to two compile time exceptions
The typenames are the only things needed for the error messages
in X::Inheritance::Unsupported and X::Composition::NotComposable.

With this change it's no longer necessary to use $! twigils when
generating the error messages (cmp. commit a822bcf9cf).

Also this makes two failing tests for RT #129906 (StackOverflowError,
probably caused by failed typechecks for the parameters in question)
in integration/error-reporting.t pass again on the JVM backend.