further enum cleanup
Enums no long backtrack on failure to find an initializer, causing
mysterious undefined enum function errors.  Fixes #67942 even though
it was closed...

Only the setting uses the special-cased enum actions for bootstrapping
purposes.  All subsequent enums just use the standard evaluator to
prevent recurring surprises that used to happen when a special case is
hit before the normal case.  Fixes #122929 and 128017.

In support of using only the happy path outside of the setting, the type
of the first pair's value can now be determined from the actual value,
not just from the AST, since we don't take the special-cased setting
path for infix:<,> in normal code anymore.

An empty initializer using a variable now warns that it was probably
not declared properly as a constant. Fixes #130446.