Fix incorrect .count-only/.bool-only impls
My undertstanding is these methods need to account for the items
already pulled from the iterator; that is, a consumer can always
do `.pull-one xx .count-only` (if .count-only is implemented) to
legally pull all of the remaining elements from the iterator.

Current impls of these methods only consider the initial state of
the iterator and do not adjust the return values as the iterator is
being pulled from.

Fix by making the methods adjust or removing the implementations.