Reify pending Seq elements before taking count
This addresses the issue of unreified but pending List and Seq elements
causing the check of reified elems to return a correct but misleading
amount the would cause eagerization to halt.

An example of such an issue is below:

    with [<a b c d e>] {
        .[0]:delete;
        with .Seq {
        use nqp;
        nqp::elems( nqp::getattr(.cache, List, '$!reified') ).say }
        }
    }

This would evaluate to 0 unless we were to explicitly reify before
taking the number of elems.

zoffixznet++ for pointing out this type of scenario.