Move attribute settage inside the lock
As far as I understand it, if we keep settage outside the lock,
there's a chance for a race where one thread would just get out of
the lock, but not yet set the attribute, while another would go into
the lock test the attribute and re-do the stuff it wasn't meant to.

Moving the attribute assignment inside the locks fixes that.