Introducing ValueObjAt, the .WHICH for value types
At this point in time, this is simply a subclass of ObjAt.  All .WHICH
methods in the core setting (apart from the fallback for Mu:D) now create
objects of this type.  With this a subclass, we gain the easy possibility
to check whether the object is a valuetype, simply by checking whether the
.WHICH ~~ ValueObjAt.  This should allow for some more optimizations in
the near future.