Make List a value type if it only has value types
- add a $!WHICH attribute to List (and therefor also on Array)
- on first call to .WHICH, will freeze the '$!WHICH' value
  - if any of the elements is a container, it's not a value type
  - if any of the elements.WHICH is not an ValueObjAt, it's not a value type
  - if all value types, create a sha of the .WHICH strings as its .WHICH
- add .WHICH for arrays that will freeze its WHICH as a non-value type ObjAt

This makes it possible to use lists that only consist of value types,
to be used in all situations where one would expect .WHICH to DWIM, such
as QuantHashes, set operators, infix:<===>, object hashes, .unique, .repeated.

This breaks two tests in roast that specifically assume that (1,2,3) === (1,2,3)
is to return False, whereas that now returns True, because the lists all
consist of just value types.

Adding an attribute to an often used object such as List/Array may induce
unnecessary overhead.  An alternative way using a mixed in Role was also
investigated, but broke a lot of assumptions on many other levels, such as
the .^name of Lists.  Even if the version of List with an $!WHICH attribute
mixed in had the same name, it would break nqp::eqaddr tests on the type
objects, which would cause problems with e.g. eqv, and thus is-deeply.

Feel free to revert this patch if consensus is that it would be too magic.