Make sure <<$a b c>> is also a value type
Originally, anything with a $!todo would be considered a non-value type.
This made <<$a b c>> not be a value type, because for some reason this
is codegenned with an iterator.  Now, we try to vivify the whole list when
a .WHICH is requested: if that fails (usually because the List is lazy),
the List is considered a non-value type.