Streamline Range.pick/roll in light of degenerate Ranges
Such as NaN..NaN or Inf..0 and stuff like that.  Any Range that cannot
determine its number of elements, will always return Nil for .pick/.roll
or an empty Seq for .pick(N)/.roll(N).  This makes the behaviour consistent
with Ranges that do not contain any elements, such as 0..^0.