guard subbuf against negative ranges
a range like 10..2 is supposed to be degenerate
and have no elements, but the code just used
int-bounds to calculate how many items to take.
This would cause attempts to allocate around 2**64
elements for the results, which is clearly bogus.

Now it'll just behave as if you had specified a
length of 0 elements.