Make `else`-less `if`/`with` blocks 3.4x faster…
…when in sunk context and when condition is false.

When `else` is absent, we stick an `Empty` in its place, to
facilitate list comprehension stuff. When in sink context,
however, it's useless and we're paying to have it sunk. Annotate
these during code gen and then turn them into nulls in the
static optimizer if we know they're gonna get sunk anyway.

Not 100% sure why the null check in the END phaser is needed.
Without it, errors about trying to call `can` on a null are raised,
presumably because we have an `END` in core with `unless { if { } }`
in it. I'm assuming these are benign and we simply don't propagate
sunkiness 100% correctly into end blocks' code receivers.
I could be wrong.