Strip Slip wrapping quoteword if one only word
When using word quoting with interpolation, interpolated words are
wrapped in a Slip. This is desirable when there are multiple words, as
the interpolated values in the Slip get integrated into the surrounding
List, but is problematic when there is only one word total (prior to
interpolation).

If there is only one word-related op in the resulting QAST tree, strip
the Slip. This fixes the issue when using both qqww and << >>.

Also added "val" sub variant that handles both Slips and Lists, as it
was there was not a good candidate for dealing with the List returned by
the "words" method (on the << >> path).

Addresses [Issue #1204](https://github.com/rakudo/rakudo/issues/1204)