Preserve operator chaining when operators negated
If the behavior on the base operator is to allow 'chaining' (as it is
for '<' and '>' in the grammar) and the metaop is '!', then preserve
chaining on the op wrapping the metaop and baseop.

Previously, applying the negation metaoperator toward operators that
allowed chaining (eg. '<', or '>') wouldn't preserve the chaining
behavior.

This commit is being made in conjunction with development in NQP
altering the chain op.

See <https://github.com/rakudo/rakudo/issues/1304>.