Fix useless useless-use warning with phasers [+Tests]
Fixes https://github.com/rakudo/rakudo/issues/1319

Since the block's return value is used to decide whether the phaser fires,
it's never useless, even if sunk. Disable warnings in those cases.

The blocks of interest appear to be already annotated with 'WANTMEPLEASE'
to force evaluation of inner loops and maps and stuff for their side effects,
when needed. So fix by annotating the last statement with `sink-quietly`
ann and then using that in places where warnings are emitted to decide
whether warning isn't required.