Fix spurious warnings with colonpaired longnames
Fixes RT#131305: https://rt.perl.org/Ticket/Display.html?id=131305

With a simple colonpair, like `infix:<+>` we have the value right
away and don't get warnings, but with things like `postfix:<[ ]>`
the value has a complex value that we compile-time compile when
figuring out what to stringify it into. When we NOT compiling core
setting, we evidently use the same machinery that makes
and—importantly—installs blocks/thunks. So with a `:<[ ]>` colonpair
in a longname, we ended up installing the Stmts with a List with two
`[` and `]` strings in it. The List was in sink context and it's that
List that was causing the spurious warnings seen in the ticket.

Fix by adding a :$no-install param to the block-making machinery and
setting it to true when compiling longname colonpairs, to avoid
installation of things we don't need.