Compile time error for misplaced whenever
I thought we already were doing this, but: make it a compile time
error to have a `whenever` block that is not lexically within a
`supply` or `react` block. Hopefully nobody was relying on this not
being an error; if they are, we can ban it in 6.d, and conditionally
do the opts that rely on it only for 6.d also. However, it seems
highly unlikely.