Optimize `supply emit <arg>`
This is a convenient way of getting a Supply that will simply emit a
single argument and then be done (or quit if the process of producing
the argument throws an exception). Given the program `my $s = supply
emit 42; for ^100_000 { react whenever $s {  } }`, this results in it
running in 78% of the time it would without this optimization.