Fix spurious useless-use warnings in metaop+metaassign op
Fixes RT#131251: https://rt.perl.org/Ticket/Display.html?id=131251

We check if the meta's op has is_pure annotation and cry if
it does, but the use isn't useless if our meta is a metassign.

N.B.: the check for metaassign can probably be moved up to the outer
conditional, but IIUC it'd make it too wide without is-from-core check.