Improve first touch of Array/Hash variables
The $!reified and $!todo of List and the $!storage of Map are lazily
allocated for variable declarations like `my @a`, `has %!h` and so
forth. However, the check to see if these fields were initialized
would allocate a Scalar due to the default auto-viv behavior of
attributes. By binding a Mu there in the prototype container that is
cloned for each created Array/Hash, we can avoid this. This also means
any code that does `my @a; ...maybe put stuff in @a... if @a { }` will
no longer allocate two Scalar containers while checking $!reified and
$!todo; similar for $!storage in a Hash.

Spotted while looking into Supply performance, and seeing Array.Bool
as the top allocator of Scalar containers!