Second-pass optimize &META_ASSIGN opt
Makes `for ^1000_000 {my $a; $a ~= "b"}` 70% faster.
In most cases, perf applies only if $a is a :U.

After we code-gen replacement for the op, we end up
with `op(nqp::p6definite($foo) ?? op() !! $foo, $bar)`.
The argless op call can often be constant-folded, so
pass our replacement through the optimizer again to
constant-fold all the things.

The NO-THROWAGE marker is there because some ops don't
have a no-arg meaning (e.g. &[%]) but we still allow
them to be used in metassign, as long as $a is a :D.
The marker tells the optimizer to now throw if it
detects that the op() will never succeeed.

The `callstatic` part of the code removed here because
it gets made in the second optimizer pass anyway (and
the pass expects `call` ops only)