Add error messages for Perl5-style pragma (e.g., "use 6.0", "need 6.0")
Fixes RT#126669: https://rt.perl.org/Public/Bug/Display.html?id=126669

When the name of an undeclared routine matches any keys of the table $obs, it shows the associated error message.
This fix leverages this table; it stores "use" and "need" each as a key and their associated error messages each as a value.