Special-case one top-level whenever supply/react
When the supply or react block ends with a single top-level whenever
(this is the case in the non-block form, but also the block form when
some setup is done followed by a `whenever`), then we can simply call
`serialize` on that Supply (which is a no-op in many cases) and that
is sufficient concurrency control. Thanks to the data structures that
are no longer needed, as well as far simpler code in message handling,
this gives a 3x-4x speedup on this relatively common case. This can
have a quite measurable impact: a Cro benchmark handles 12% more
requests per second thanks to this (without any modifications made to
Cro itself).