Remove src/core/UInt64.pm
This appears to be a leftover from at about 7 years ago.  It created a
subset that was not being used anywhere and a constant that was needed
for Rat/FatRat.  UInt64 (not to be confused with uint64) is not documented
anywhere and not tested.  Moved the constant to src/core/Rat and changed
it into a real constant.

Kudos to Andrew Shitov for spotting this:
  https://perl6.online/2018/01/15/26-native-integers-and-uint-in-perl-6/