Fix crash with colonpaired variables declarations
Fixes RT#128712: https://rt.perl.org/Ticket/Display.html?id=128712

When we reference a variable inside the colonpair, the declaration
marker from the variable we're declaring is still set, so the lookup
inside the colonpair explodes because scope is not set. Fix by
resetting that marker when we start parsing the colonpair circumfix.