Avoid unnecessary check of PrecompilationId's validity
In many cases we can trust that the precompilation id is already valid,
because it's derived from the compiler's id or is generated by way of sha1
or comes from the name of a file named after a precompilation id.
So avoid the costly regex parsing of the id in those cases and just trust
the value.