Fix `where` optimization for Junction args
Phixes https://github.com/rakudo/rakudo/issues/1457

Our optimization uses `istype` op, which doesn't thread, so with
unthreaded params, instead of doing a smartmatch with a Junction,
we end up doing istype on it and blowing up.

Fix by annotating unthreaded params and using a slightly more
elaborate optimization for them, where we check if the arg is a
Junction and use the slower-path ACCEPTS call in those cases.