Fix crash in Proxy.perl
Phixes https://github.com/rakudo/rakudo/issues/1466

The crash occurs due to `self` of a Proxy-ed object being conted,
but access to it fetches the value. So we end up entering .perl
via Mu:D candidate (the Proxy object itself) and then calling
.perlseen method on it (thus on the fetched value). So if that value
is a :U, then we explode, because there ain't no :U .perlseen candidate.
If the value is a :D, we'd generate incorrect .perl for many types.

Fix by checking if `self` is conted to discern when we're dealing
with a Proxy and then calling .perl on the fetched value. Not 100% sure
it should behave like that, rather than giving some sort of
"Proxy.new: …" value.