Elevate statevars to enclosing block when wrapped
Previously in blocks following the "do" statement prefix, statevars
would be seemingly reset after each iteration of the loop.

Given the statement:

    say do loop { state $x = 0; last if ++$x > 2; $x }

This statement would result in multiple thunk refs being produced, and
ultimately, would be wrapped in a block that is used as a cloning target
in looping and a new container would be declared and initialized for
each iteration.

The QAST would look like this (trimmed to relevant parts):

    - QAST::Block(:cuid(3) :blocktype(declaration_static))
      - QAST::Stmts  { state $x = 0; last if ++$x > 2; $x }
      - QAST::Stmts
        - QAST::Block(:cuid(1) :blocktype(immediate)) <wanted>
          - QAST::Stmts
            - QAST::Var(lexical $x :decl(statevar))
            - QAST::Var(lexical $_ :decl(var))

After this commit the QAST looks like this:

    - QAST::Block(:cuid(3) :blocktype(declaration_static))
      - QAST::Stmts  { state $x = 0; last if ++$x > 2; $x }
        - QAST::Var(lexical $x :decl(statevar))
      - QAST::Stmts
        - QAST::Block(:cuid(1) :blocktype(immediate)) <wanted>
          - QAST::Stmts
            - QAST::Op(null)
            - QAST::Var(lexical $_ :decl(var))

This way the statevar is declared in the block being cloned and thus the
same container is used in each iteration.

See [Github Issue 1341](https://github.com/rakudo/rakudo/issues/1341)