Fix broken 2-arity sort of 2-el list
Phixes https://github.com/rakudo/rakudo/issues/1472

When we have a 2-el list, we need to swap only if Order::More
is received from the comparator, but the test was checking
whether to leave it alone if Order::Less is received.

Fix by leaving it alone for Order::Same as well.