Re-design Whatever currier
- Fixes RT#128859: https://rt.perl.org/Ticket/Display.html?id=128859
- Makes (* + * + * + * + *)(1,2,3,4,5) run 2.2x faster
- Makes (* + * + * + * + *)(1,2,3,4,5) compile ~8% faster

=== [The mechanism of the OLD currier] ===

We go over the QAST, replacing the Whatever and HyperWhatever with
vars that get their values from params of our curry. Whenever we
encounter a WhateverCode, we simply add more whatever-params to
our curry and invoke the WhateverCode with those variables. The
result is a simple expression that takes 5 curried args, like:

    (* + * + * + * + *)(1,2,3,4,5)

Compiles to a tree of 4 Callables that make 4 additional
infix:<+> calls and that use 14 variables to do the job.

For chained calls, the assumption is made that a WhateverCode inside
a chained arg is made from earlier op in the chain. This assumption
is incorrect and is the cause of RT#127759, where the WhateverCode
in the chain is a result of currying a method call.

Otherwise, chained calls are handled by using 4 annotations holding
variables, params, the name of the chained op, and curry QAST::Block.
When the next link of the chain is made, these annotations are raked
for goods that are moved into the new link and the old QAST::Block
gets disabled by having its guts replaced by a die call.

=== [The mechanism of the NEW currier] ===

The new currier creates a "flat" structure with a single Callable,
whenever we encounter a WhateverCode, we move its guts to our
current curry and disable that block. Just like the old currier
did for chained calls.

Instead of shoving everything into annotations, we gut the
WhateverCode's past_block annotation, ignoring any :decl<param>,
vars, as those are from the signature maker, from when we curried
that WhateverCode. (we clone the QAST::Vars in the declarations part,
as the static optimizer's lowering messes stuff up if we don't).

The statement portion of the gutted QAST::Block replaces the
WhateverCode call in our current curry, and since we copied the
whatever variables into the statement portion (and added them to
the signature maker for current curry), we don't need to modify
any names or anything else in the statement portion.