Separate adding child block from creating code obj
Previously, the create_simple_code_object method in the Perl6 World
would add the block passed to it to the desired outer block in addition
to returning a code object for it. From the name, it is not clear that
the method has such side-effects. Separating the actions in order to
make it explicit that we performing both of these.