Fix drift when reusing lazy iterables for indexing
Phixes https://github.com/rakudo/rakudo/issues/1320

Currently, we .push-until-lazy before we begin checking whether
for existence of elements to know when to stop. However, this
causes a "drift" when the lazy iterable is reused as .push-until-lazy
reifies an extra element each time to figure out if it's still lazy.
The next time we re-use this iterable for indexing, that extra element
ends up being past end of the iterable we're indexing; with an extra
such element added on each reuse.

Fix by checking if iterable is lazy and checking all positions for
existence for such iterables each time.