Fix Any.tail(Callable) violating Iterator protocol
Fixes RT#131617: https://rt.perl.org/Ticket/Display.html?id=131617

When we .skip-at-least on our iterator, it's possible we skipped
all the values already. When we then use that same iterator
to make a new Seq, we violate protocol when that Seq is consumed.

Fix by checking return value of .skip-at-least and returning
an empty Seq when we skipped all the values.