Fix multi-threaded precompilation
The lock file is sufficient to protect the precompilation store from access
by multiple processes. However if multiple threads trigger
precompilation they will use the same store object which thinks its
already sufficiently locked. So multiple writers within the same process
will access the store concurrently. We cannot rely on the lock file in
this case, since it will be the same file handle in all threads and the
.lock call does not block if the file handle is already locked in the
same process.

Use the $!update-lock instead to serialize access to a precompilation
store within a process. This fixes odd errors when multiple threads
trigger precompilation of modules with a shared dependency concurrently.

Thanks to ugexe++ for the report and the initial diagnosis.