Fix potential issue in Sprintf exception handler
First, in the code we expect to throw the return value of
R::I.HANDLE-NQP-SPRINTF-ERRORS, but it already throws. Second,
if the payload doesn't mention any of the keys, we'll end up
returning an `Empty` and trying to call `.throw` on it.

Fix by having R::I.HNSE return an Exception object: either the
updated sprintf exception or the original one, and then we'll throw
it at the point of call.