Add "Is HLL Init" to show if statevar is assigned
Given the following subroutine:

    sub f($x) {
        return if $x;
        state $y = 5;
        $y;
    }

Previously, $y would only be assigned a value if the statement is
reached on the first invocation of the closure.

To address this, added an indicator on the code ref associated with the
block (in corresponding MoarVM development) and modified p6stateinit to
take a var and check whether it has been assigned to. If the assignment
was not reached in the first pass in the life of a closure, the
is_hll_init value indicates that it requires HLL assignment the next
time it is encountered.

See [RT#102994](https://rt.perl.org/Public/Bug/Display.html?id=102994)