Make Backtrace.gist more helpful
Fixes R#1521 https://github.com/rakudo/rakudo/issues/1521

Currently we go through Mu.gist, which is just Mu.perl, which
just gives "Backtrace.new", leading users to believe it's an empty
backtrace.

Implement Backtrace.gist that prints "Backtrace(42 frames)".
Reasoning for not making .gist do .Str is because .Str is formatted
like a typical error dump, so if you .gist, say, a list of Backtaces,
you get weird-looking results. This way users have an option of
getting a "gist" of a backtrace as well as full frame dump with
`print` instead of `say`.