Differentiate precomp NC sub setup markers
Fixes https://github.com/rakudo/rakudo/issues/1576 R#1576

If a NativeCalled sub used is during precomp, it's marked as
`$!setup`. However, when it's later used in non-precomp environment,
the stuff setup does is not available and so we get a SEGV.

Fix by using a separate marker for $!setup when we're in precomp
mode so that when we switch to non-precomp mode, we setup again.