Make infix:<≠> an alias of infix:«!=»
By making it an alias, we can get rid of all of the duplication for the
unicode operator.  This reduces memory size (about 3K in the setting.moarvm)
and maintenance burden (because additional candidates will only need to be
added for the ascii version of the operator).

Please note that the alias *must* use <> brackets: apparently creating
constants takes literal names, and 'sub infix:«≠»; normalizes to <> brackets.