Further nano-optimize threadpool scheduler thread
It struck me that we don't need the overhead of calling the HLL sleep()
as we know the value that will be used all of the time.  Please note that
I also tried to feed it a native num variable, but this appeared to have
a negative effect on longer running parallel executing programs.