Populate Signal enum with getsignals op
Previously, Rakudo and it's backends used system-independent signal
values and transformed them to the proper value when submitting a signal
handler. However when used in conjunction with NativeCall and the native
C kill implementation, the fact that the Signal values do not match the
host system becomes a problem.

Making it so that the signal numbers exposed to the user as Signal enums
will match those of their host system. They can be set via the
getsignals op, which will retrieve the valid signals from the host.