Fix .grep(Regex) on Hyper/Race Seqs
Regex is a Callable, but we can't just invoke it with any
object we want and gotta call .ACCEPS, which will set up
the cursor and invoke the regex with it.

Fix by unifying GrepSM and GrepCode classes. Handling the Regex case
in GrepCode makes that case 99% the same as GrepSM, so just unify
them and make the class clone Callables and invoke them directly,
while Regexes and rest of stuff would get .ACCEPTS calls.