Make MONKEY-TYPING imply `no precompilation`
It's not possible to load two precompiled modules that both do a
monkey-patch to the same type. Precompiling them will thus result in
modules that cannot be loaded together with others.

So far we've expected module authors to read the docs and remember to
add `no precompilation`. That's less reliable than the compiler just
doing the right thing for them. Especially given that the problem will
probably never be noticed by a module author when writing tests for
their module, but instead when somebody uses that module together with
others in a larger application.