Streamline Channel.receive/receive-nil-on-close
There is no difference between these two methods apart from the return
value when the Channel closes.  Therefore, it was implemented with a
private method taking a flag.

However, when you use .hyper on something longer running, the profile
shows that Channel!receive is not even speshed.  Therefore it made sense
to streamline and duplicate the .receive logic, so that it can be speshed
and JITted.

This makes test-t --race <hello20.csv about 2.5% faster, and uses about
2.8% less CPU.