Streamline .hyper.grep batch producer
This makes the producer go from 810 msecs to 639 msecs for something like
  say ^Inf .hyper.grep( *.is-prime ).skip(9999).head

Instead of creating a new IterationBuffer for the result, the existing
IterationBuffer is reused.

Oddly enough, this appears to have a *negative* effect on wallclock and
CPU usage overall.  Perhaps because this is reducing workload in the threads,
so workers finish more quickly and start fighting for the next job.

The --profile shows significantly fewer allocations and 2 extra OSR's.  So it
feels worthwhile to keep.