Fix precomp files of NativeCall users containing absolute paths
By storing the compiler object in a variable scoped to the Native role, it
gets serialized into precomp files. This compiler object contains a lot of
information including the path to the source file of the currently compiling
module (the NativeCall user) and the path of the target bytecode file.

The precomp files created this way fail linter tests used by openSUSE on the
Open Build Service which prohibit storing BUILDROOT paths in binary files -
a very sensible restriction.

Fix by moving the variable into the only function that actually uses it.