Don't use typevar inappropriately
It instructs optimizers that if they are holding the type of the first
argument (in a method, "the invocant") constant, then the lookup of the
lexical can also be assumed constant also. However, we were using it in
nested blocks, where a first argument type could be held constant but it
was totally unrelated to any notion of "invocant".

This adds a safety check, falling back to the safe `lexical` scope if it
sees that the symbol is not declared in the package immediately around
the current method. This will rule out some cases that are safe too (like
when the optimizer flattens scopes); we could restore those later.