make sure MERGESORT-* don't leak low-level types
depending on how many items were in the list, you
could end up getting a BOOTIntArray instead of an
intarray, for example with 65 elements.

This was caused by `$A` and `$B` switching places after
every "round" of mergesort application, and if you
got the right number of rounds, you'd get `$B` instead,
which was just a nqp::list_i (or _n or _s).