Support the "api" adverb when loading modules
:api is semver's major version, i.e. it signifies a backards incompatible
change. So whenever a module author changes the module's API in a backwards
incompatible way (i.e. not just an extension), changing the value of :api
gives a signal to the user. This way on the receiving end, one can specify
"give me the newest version of the module, but stay in the range that still
provides the API I expect" independent of the nuances of versions.

In other words: we treat modules with different values for :api as essentially
different modules while we assume that just any higher version than the minimum
specified will be OK.