Make builds reproducible
We replace the compilation timestamp by a hash of the complete source (both
the setting and the metamodel), so a recompilation without changes will give
the exactly same result.
This helps distro packagers as it makes the installation path predictable
and removes the requirement to re-build all modules after a re-build of rakudo.