Consistify &combinations/&permuations with method forms
In most of the language, subroutine forms are largely a functional
way to use the method forms. However, with &combinations/&permutations,
trying to follow that convention produces confusing[^1] results in that
@list.permuations permutates the list, yet permutations(@list) is
equivalent to permutations(^@list). If @list contains small integers,
this issue can be hard to spot too.

This commit makes the two subs more consistent with their method forms:
- Add permutations(Iterable \n) and combinations(Iterable \n, \k)
    candidates that work like n.permutations/n.combinations: k
- Make combinations(\n, Range) work like it does in method form
    (previously treated Range as Int)
- Make combinations(\n) work like it does in method form
    (previously threw X::Multi::NoMatch)

[1] https://www.reddit.com/r/perl6/comments/84vrhk/adjective_noun_pick_and_choose_part_2/dvvccou/