Replace HyperToIterator.push-all by push-exactly
The difference between the two is just one int condition and one int assignment
per HyperWorkBatch, which is neatly shown in the diff of this commit.
The extra work, at least currently, completely gets lost in the noise.

Also fix superfluous check and return value in skip-at-least: we only exit the
loop through -return- anyway.