Make .produce/.reduce nodal the correct way
The intent seems to be that they should be nodal, but they way this
was implemented, hid the "is nodal" functionality.

Simplify the actually working candidate to not need a temporary variable
or an -if- statement: if the lookup for the reducer fails, a Failure will
be returned, which will throw as soon as it is attempted to be executed.

Also implement a candidate for handling type objects: there should be
no reason for doing that at runtime repeatedly.

This breaks tests 134/135 in S03-metaops/hyper.t .  However, these tests
seem to assume non-nodality.

This may also have breaking consequences in the ecosystem: a special note
in the release message / a p6alert seems to be in order if this change is
going to make it to the next release.