Make a HyperBatcher responsible for its own sequencing
Some HyperBatchers may not need a sequence number as part of the Batcher.
Having it in a role, and currently not being able to use writable native
int attributes from a role, made the sequence number fetching appear quite
high in profiles.  This was fixed a while ago in the default HyperIteratorBatcher
by giving it its own native int sequence number.