Move parameter checking to HyperConfiguration.new
Whether :batch and :degree are valid or not, is really best checked
inside the HyperConfiguration class itself.  Initially I moved the
Iterable!valid-hyper-race to HyperConfiguration, but then it felt
contrived.  So bit the bullet and moved the check to a TWEAK submethod.

This should make it easier to create .hyper / .race methods in other
classes, such as Range, Array, List, etc.